LPS-57749 renames PortletProviders to follow the <portletName>-<action>-PortletProvider pattern
