--- modules/apps/users-admin/users-admin-web/src/com/liferay/users/admin/web/portlet/OrganizationBrowsePortletProvider.java
@@ -1,55 +1,0 @@
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
-
-package com.liferay.users.admin.web.portlet;
-
-import com.liferay.portal.kernel.exception.PortalException;
-import com.liferay.portal.kernel.portlet.BasePortletProvider;
-import com.liferay.portal.kernel.portlet.BrowsePortletProvider;
-import com.liferay.users.admin.web.constants.UsersAdminPortletKeys;
-
-import javax.portlet.PortletURL;
-
-import javax.servlet.http.HttpServletRequest;
-
-import org.osgi.service.component.annotations.Component;
-
-/**
- * @author Pei-Jung Lan
- */
-@Component(
-	immediate = true,
-	property = {"model.class.name=com.liferay.portal.model.Organization"},
-	service = BrowsePortletProvider.class
-)
-public class OrganizationBrowsePortletProvider
-	extends BasePortletProvider implements BrowsePortletProvider {
-
-	@Override
-	public String getPortletId() {
-		return UsersAdminPortletKeys.USERS_ADMIN;
-	}
-
-	@Override
-	public PortletURL getPortletURL(HttpServletRequest request)
-		throws PortalException {
-
-		PortletURL portletURL = super.getPortletURL(request);
-
-		portletURL.setParameter("mvcPath", "/select_organization.jsp");
-
-		return portletURL;
-	}
-
-}
